Added new function
Improve the method to show the downloaded files at its directory.

The previous method was to call the command `Run %ComSpec% /c %dl_cmd% && explorer %download_dir% && exit`

Now there's a dedicated function check_explorer_path(path), to check if the explorer.exe is already opened and update it or open a new window if it is not.

Also included project_notes.md to help me track this project.